ai/requirements.txt
opencv-python==4.9.0.80
face-recognition==1.3.0
websocket-client==1.7.0
numpy==1.26.3
python-dotenv==1.0.0
Pillow==10.2.0
